+++ src/cli/command-registrar-builder.js
@@ -38,6 +38,7 @@ import Remove from './commands/public-cmds/remove-cmd';
 import Deprecate from './commands/public-cmds/deprecate-cmd';
 import DeprecatePrivate from './commands/private-cmds/_deprecate-cmd';
 import Delete from './commands/private-cmds/_delete-cmd';
+import Latest from './commands/private-cmds/_latest-cmd';
 
 export default function registerCommands(): CommandRegistrar {
   return new CommandRegistrar(BIT_USAGE, BIT_DESCRIPTION, BIT_VERSION, [
@@ -77,6 +78,7 @@ export default function registerCommands(): CommandRegistrar {
     new Remove(),
     new Deprecate(),
     new Delete(),
-    new DeprecatePrivate()
+    new DeprecatePrivate(),
+    new Latest()
   ]);
 }