+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - prevent bare-scope corruption when the export process fails 
 - fixed stderr maxBuffer exceeded bug in ci-update cmd
+- add `--outdated` flag to `bit list` command to show the remote versions of components
 
 ## [0.10.9] - 2017-10-18